refactor(commands): extract shared UI settings helpers into ui_settings.rs
